pyproject.toml
#  Copyright 2024 SkyAPM org
#
#  Licensed under the Apache License, Version 2.0 (the "License");
#  you may not use this file except in compliance with the License.
#  You may obtain a copy of the License at
#
#  http://www.apache.org/licenses/LICENSE-2.0
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.

[tool.poetry]
name = "R3"
version = "0.1.0"
description = "A URI Endpoint Grouping Algorithm for SkyWalking AI Pipelines"
license = "Apache-2.0"
authors = ["Superskyyy <yihaochen@apache.org>"]
readme = "README.md"
homepage = "https://skywalking.apache.org/"
repository = "https://github.com/SkyAPM/R3"
keywords = ["skywalking", "observability", "aiops", "machine-learning", "URI"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: System :: Monitoring",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
]


packages = [
    { include = "servers" },
    { include = "models" }
]

# poetry will ignore generated files as .gitignore, but we need them in package
include = ['servers/protos/**/*']
exclude = ['tests']


[tool.poetry.urls]
"Issues Tracker" = "https://github.com/apache/skywalking/issues"
"Documentation" = "TBD"

[tool.poetry.scripts]
# entry point for the CLI

[tool.poetry.build]
generate-setup-file = true

[tool.poetry.dependencies]
python = ">=3.8,<3.12"  # Gradio needs 3.8+
grpcio = "*"
grpcio-tools = "*"
jsonpickle = "^3.0.1"
cachetools = "^5.3.1"
nltk = "^3.8.1"
inflect = "^6.0.4"
pytest = "^7.3.2"
apache-skywalking = "^1.0.1"
flask = "^2.3.2"




[tool.poetry.dev-dependencies]
#PySnooper = "^1.1.1"
## TESTING
#tox = "^3.25.1"
#tqdm = "^4.64.0"
#yappi = "^1.3.6"


[tool.poetry.group.lint.dependencies]
pylint = '2.13.9'
flake8 = "^5.0.4"
# isort = "^5.10.1"
unify = "^0.5"
flynt = "^0.76"
# flake8-quotes = "^3.3.1"
flake8-use-fstring = "^1.4"
pep8-naming = "^0.13.2"
darglint = "^1.8.1"
flake8-eradicate = "^1.4.0"
flake8-docstrings = "^1.6.0"
flake8-bugbear = "^22.9.23"
flake8-comprehensions = "^3.10.0"
flake8-2020 = "^1.7.0"


[tool.poetry.group.demo.dependencies]
gradio = "^3.35.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"